Account Class Completed

+++ OOP2_Mid_Term_Assignment/OOP2_Mid_Term_Assignment/Bank.cs
@@ -42,7 +42,7 @@ namespace OOP2_Mid_Term_Assignment
             }
 
         }
-        public void Transaction (int choice, int amount, int accountNumber, Account receiver)
+        public void Transaction (int choice, double amount, int accountNumber, Account receiver)
         {
             switch(choice)
             {
